+++ code/simulator_yaml_parser/src/force_parsers.cpp
@@ -42,12 +42,12 @@ Eigen::Matrix<double,6,6> parse_quadratic_damping(const std::string& yaml)
     parser.GetNextDocument(node);
     YamlDynamics6x6Matrix M;
     node["damping matrix at the center of gravity projected in the body frame"] >> M;
-    for (size_t j = 0 ; j < 6 ; ++j) ret(0,j) = M.row_1[j];
-    for (size_t j = 0 ; j < 6 ; ++j) ret(1,j) = M.row_2[j];
-    for (size_t j = 0 ; j < 6 ; ++j) ret(2,j) = M.row_3[j];
-    for (size_t j = 0 ; j < 6 ; ++j) ret(3,j) = M.row_4[j];
-    for (size_t j = 0 ; j < 6 ; ++j) ret(4,j) = M.row_5[j];
-    for (size_t j = 0 ; j < 6 ; ++j) ret(5,j) = M.row_6[j];
+    for (size_t j = 0 ; j < 6 ; ++j) ret(0,(int)j) = M.row_1[j];
+    for (size_t j = 0 ; j < 6 ; ++j) ret(1,(int)j) = M.row_2[j];
+    for (size_t j = 0 ; j < 6 ; ++j) ret(2,(int)j) = M.row_3[j];
+    for (size_t j = 0 ; j < 6 ; ++j) ret(3,(int)j) = M.row_4[j];
+    for (size_t j = 0 ; j < 6 ; ++j) ret(4,(int)j) = M.row_5[j];
+    for (size_t j = 0 ; j < 6 ; ++j) ret(5,(int)j) = M.row_6[j];
     return ret;
 }
